backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
pdfplumber==0.11.4
python-multipart==0.0.20
openai>=1.0.0
pydantic==2.10.4
python-dotenv>=1.0.0
python-telegram-bot>=22.0
httpx>=0.27.0
